Refactor native components queries and typings
