+++ react/components/form/BrandSelector/types/typings.d.ts
@@ -1,0 +1,8 @@
+export interface Brand {
+  name: string
+  id: string
+}
+
+export interface Data {
+  brandSearch: Brand[]
+}
