+++ react/components/form/CategorySelector/queries/Categories.graphql
