+++ react/components/form/CategorySelector/types/typings.d.ts
@@ -1,0 +1,8 @@
+export interface Category {
+  name: string
+  id: string
+}
+
+export interface Data {
+  categorySearch: Category[]
+}
