+++ react/components/form/CollectionSelector/queries/Collections.graphql
