+++ react/components/form/CollectionSelector/types/typings.d.ts
@@ -1,0 +1,8 @@
+export interface Collection {
+  name: string
+  id: string
+}
+
+export interface Data {
+  collectionSearch: Collection[]
+}
