+++ react/components/form/DepartmentSelector/types/typings.d.ts
@@ -1,0 +1,8 @@
+export interface Department {
+  name: string
+  id: string
+}
+
+export interface Data {
+  departmentSearch: Department[]
+}
