+++ react/components/form/typings.d.ts
@@ -1,0 +1,11 @@
+export interface Option {
+  label: string
+  value: any
+}
+
+export interface State <T> {
+  data: T
+  errors: any
+  loading: boolean
+  value: any
+}
